traitement des retours et nettoyage de code [APPS-01RW]
